Matchup Analysis in the Evaluation Function
Using the `Checks and Counters` section of the Smogon Stats JSON we can
get an idea of which of the two active pokemon has an advantage.

The data is statistics-based and isn't perfect. Which of the two active pokemon
has the advantage is more than just the matchup - other factors are HP, statuses,
volatile-statuses, and more. However empirically I have
found that the bot performs at an average of ~+100 elo and ~+4% GXE when
this is used versus when it is not.